+++ test/base/src/androidTest/kotlin/io/realm/test/shared/MutableRealmTests.kt
@@ -18,13 +18,14 @@ package io.realm.test.shared
 import io.realm.MutableRealm
 import io.realm.Realm
 import io.realm.RealmConfiguration
-import io.realm.RealmInstant
-import io.realm.delete
-import io.realm.entities.SampleWithPrimaryKey
+import io.realm.RealmResults
+import io.realm.entities.Sample
 import io.realm.entities.StringPropertyWithPrimaryKey
 import io.realm.entities.link.Child
 import io.realm.entities.link.Parent
 import io.realm.query
+import io.realm.query.RealmQuery
+import io.realm.query.RealmSingleQuery
 import io.realm.test.platform.PlatformUtils
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.runBlocking
@@ -34,6 +35,7 @@ import kotlin.test.Ignore
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
+import kotlin.test.assertNotEquals
 import kotlin.test.assertNotNull
 import kotlin.test.assertNull
 import kotlin.test.assertTrue
@@ -53,6 +55,7 @@ class MutableRealmTests {
                 Parent::class,
                 Child::class,
                 StringPropertyWithPrimaryKey::class,
+                Sample::class,
                 SampleWithPrimaryKey::class
             )
         ).path("$tmpDir/default.realm").build()
@@ -459,12 +462,86 @@ class MutableRealmTests {
     }
 
     @Test
-    fun delete() {
+    fun delete_realmObject() {
         realm.writeBlocking {
             val liveObject = copyToRealm(Parent())
             assertEquals(1, query<Parent>().count().find())
             delete(liveObject)
             assertEquals(0, query<Parent>().count().find())
+        }
+    }
+
+    @Test
+    fun delete_realmList() {
+        realm.writeBlocking {
+            val liveObject = copyToRealm(Sample()).apply {
+                stringField = "PARENT"
+                objectListField.add(Sample())
+                objectListField.add(Sample())
+                objectListField.add(Sample())
+                stringListField.add("ELEMENT1")
+                stringListField.add("ELEMENT2")
+            }
+
+            assertEquals(4, query<Sample>().count().find())
+            assertEquals(3, liveObject.objectListField.size)
+            assertEquals(2, liveObject.stringListField.size)
+            delete(liveObject.objectListField)
+            delete(liveObject.stringListField)
+            assertEquals(0, liveObject.objectListField.size)
+            assertEquals(0, liveObject.stringListField.size)
+            assertEquals(1, query<Sample>().count().find())
+        }
+    }
+
+    @Test
+    fun delete_realmQuery() {
+        realm.writeBlocking {
+            for (i in 0..9) {
+                copyToRealm(Sample().apply { intField = i % 2 })
+            }
+            assertEquals(10, query<Sample>().count().find())
+            val deleteable: RealmQuery<Sample> = query<Sample>("intField = 1")
+            delete(deleteable)
+            val samples: RealmResults<Sample> = query<Sample>().find()
+            assertEquals(5, samples.size)
+            for (sample in samples) {
+                assertEquals(0, sample.intField)
+            }
+        }
+    }
+
+    @Test
+    fun delete_realmSingleQuery() {
+        realm.writeBlocking {
+            for (i in 0..3) {
+                copyToRealm(Sample().apply { intField = i })
+            }
+            assertEquals(4, query<Sample>().count().find())
+            val singleQuery: RealmSingleQuery<Sample> = query<Sample>("intField = 1").first()
+            delete(singleQuery)
+            val samples: RealmResults<Sample> = query<Sample>().find()
+            assertEquals(3, samples.size)
+            for (sample in samples) {
+                assertNotEquals(1, sample.intField)
+            }
+        }
+    }
+
+    @Test
+    fun delete_realmResult() {
+        realm.writeBlocking {
+            for (i in 0..9) {
+                copyToRealm(Sample().apply { intField = i % 2 })
+            }
+            assertEquals(10, query<Sample>().count().find())
+            val deleteable: RealmResults<Sample> = query<Sample>("intField = 1").find()
+            delete(deleteable)
+            val samples: RealmResults<Sample> = query<Sample>().find()
+            assertEquals(5, samples.size)
+            for (sample in samples) {
+                assertEquals(0, sample.intField)
+            }
         }
     }